Use bash rather than sh
Update shebang line to always use bash, even on systems where sh is
not a symlink to that shell.

At least some of the syntax in wrye-install relies on non-POSIX shell
extensions in bash.  On some Linux systems, sh is symlinked to
/bin/bash.  On others, it is not.  On current Debian-family Linux
systems, it is symlinked to /bin/dash, which does not understand the
"[[" bash extension, which was causing wrye-install to fail with
syntax errors on such systems.